+++ examples/hello_world/src/main.rs
@@ -1,0 +1,7 @@
+fn main() {
+    for i in 0..8 {
+        let s = format!("Hello, {}", i);
+        println!("{}", s);
+    }
+    panic!("OK");
+}